requirements.txt
google-api-python-client
google-cloud-secret-manager
anthropic
flask
requests
python-dotenv